% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{pm25_india}
\alias{pm25_india}
\title{Table of PM2.5 concentrations in 5 cities 2013-2016}
\format{A tbl_df (dplyr) with a \code{POSIXct} datetime column, a \code{character} column with the city name and the PM2.5 concentration in microg/m3 as a \code{numeric} column.}
\description{
Table of PM2.5 concentrations in 5 cities 2013-2016: Delhi, Chennai, Kolkata, Hyderabad, Mumbai. Please refer to
original source (U.S. Embassy and Consulates General in India air quality monitoring program) \url{http://newdelhi.usembassy.gov/airqualitydata.html} for licensing details.
}
\examples{
data(pm25_india)
str(pm25_india)
}
\keyword{data}
