% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{usaqmindia_calendar}
\alias{usaqmindia_calendar}
\title{Calendar plot for one city}
\usage{
usaqmindia_calendar(cityplot = NULL)
}
\arguments{
\item{cityplot}{the city, "Delhi", "Chennai", "Kolkata", "Hyderabad", "Mumbai".}
}
\description{
Calendar plot for one city
}
\examples{
usaqmindia_calendar(cityplot = "Chennai")
}

