% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redo_plot.R
\name{usaqmindia_plot}
\alias{usaqmindia_plot}
\title{Re-do the time series plot for all cities}
\usage{
usaqmindia_plot()
}
\value{
a ggplot graph
}
\description{
Re-do the time series plot for all cities
}
\examples{
usaqmindia_plot()
}

